0x0D. C - Preprocessor
TASKS:

0. Object-like Macro

Create a header file that defines a macro named SIZE as an abbreviation for the token 1024.

1. Pi

Create a header file that defines a macro named PI as an abbreviation for the token 3.14159265359.

2. File name

Write a program that prints the name of the file it was compiled from, followed by a new line.

You are allowed to use the standard library

3. Function-like macro

Write a function-like macro ABS(x) that computes the absolute value ofa number x.

4. SUM

Write a function-like macro SUM(x, y) that computes the sum of the numbers x and y.

q